Create rgba.html
This HTML code showcases the use of RGBA (Red, Green, Blue, Alpha) color values to style elements. The <h2> element has a solid gray background (rgba(128, 128, 128, 1.0)), and a <table> has a semi-transparent red background (rgba(255, 0, 0, 0.8)). Inside the table, a <p> element is meant to have white text (rgba(255, 255, 255, 1.0)), but there's a typo: it incorrectly uses gba instead of rgba, which will result in the color not being applied. Fixing this typo will correctly style the paragraph text with white color.