Create iframe size.html
This HTML code demonstrates how to set the height and width of an iframe. The <iframe> element is used to embed content from /index.htm with a specified width of 500 pixels and a height of 300 pixels. If the browser does not support inline frames, the text "Sorry your browser does not support inline frames" will be displayed as a fallback message. This is a simple example of how to embed external content while controlling the size of the iframe.