Create meta tags.html
This HTML code demonstrates the use of meta tags in the <head> section of a webpage. The meta tags provide metadata about the webpage, which is used by search engines and browsers:

<meta name="keywords" content="HTML, Meta Tags, Metadata" />: Specifies keywords related to the page's content for search engines.

<meta name="description" content="Learning about Meta Tags." />: Provides a brief description of the page, which can be shown in search engine results.

<meta name="revised" content="hello, 3/7/2014" />: Indicates when the page was last revised.

The body of the page contains two paragraphs: one with the text "Hello HTML5!" and another showing the date "3/7/2014". Meta tags do not display on the webpage itself but serve as important information for web crawlers and browsers.