Create del tag.html
In this HTML code:

The <del> tag is used to indicate that the word "Madras" has been deleted or is no longer relevant. It typically displays the text with a strikethrough effect.

The <ins> tag is used to indicate that the word "Chennai" has been inserted or added. This usually displays the text with an underline to highlight the addition.

So, in the paragraph:

"Madras" will appear with a strikethrough.

"Chennai" will appear underlined.

This is useful for showing changes or updates in content, such as when editing or revising a document.