Create font face attribute tag.html
This HTML code demonstrates the use of the <font> tag, which is used to specify the font style and size of text:

<font> tag: The <font> tag is used to specify the font face (typeface) and size of the text within the tag. It has been deprecated in HTML5 and is no longer recommended for use in modern web development. Instead, CSS should be used for styling.

Attributes:

face: Specifies the font family to be used for the text. Examples include Times New Roman, Verdana, Comic sans MS, etc.

size: Specifies the size of the text. The number 5 is used here, which corresponds to a medium size according to HTML standards (this varies depending on the browser).

Text Variations:

The text for each font is displayed with different font families, showing examples of commonly used fonts.

Some fonts may not display properly if they are not installed on the system, like "WildWest" or "Bedrock" which are custom fonts not commonly found on all devices.

The output will show the text "Times New Roman," "Verdana," "Comic Sans MS," "WildWest," and "Bedrock" in their respective font styles and size 5.