Create font size attribute tag.html
This HTML code demonstrates the use of the <font> tag to change the font size of the text:

<font size="x">: The <font> tag is used here with the size attribute to specify the size of the text. The size attribute accepts values from 1 to 7, where 1 is the smallest size and 7 is the largest size in the context of the HTML <font> tag.

Font Size Values:

size="1": This will render the text with the smallest size.

size="2": This will render the text slightly larger.

size="3": This renders the text at a medium size.

size="4": Larger text than size 3.

size="5": This renders the text at an even larger size.

size="6": Larger than 5.

size="7": This will render the text with the largest available size.

The output will show the text "Font size = X" where X is the respective font size. For example, "Font size = 1" will appear the smallest, while "Font size = 7" will appear the largest.

Note: The <font> tag is deprecated in HTML5 and should be avoided in modern web development. It is recommended to use CSS for styling font sizes.