Create drop-down feature of a form.html
This HTML code creates a form containing a select box (also known as a dropdown menu) with the name dropdown. The menu provides three options: Maths, Physics, and Chemistry, where Maths is preselected by default using the selected attribute. When a user chooses a subject, the corresponding value (e.g., "Maths") can be submitted with the form.