Create id4(Simple ID Attribute).html
This HTML document demonstrates the use of the id attribute along with CSS to style specific <div> elements. The first <div> has the ID TP and displays the text “TutorialsPoint Website” with dark blue color (#070770), bold font weight, and a font size of 30px. The second <div> has the ID TP1 and displays the text “Html5 Tutorials, CSS Tutorials, JavaScript Tutorials” with a font size of 25px. However, there's a small typo in the CSS: the value of text-align should be "center" instead of "centre" for proper alignment in CSS.