Create class1.html
This HTML document uses the highlight class to style the <p> element with a yellow background, black text, and padding. However, the <h1> element is using a class named header, which does not have any associated styles in the provided CSS. If the intention was to apply similar styling to the <h1> element, the header class would need to be defined in the CSS. The highlight class works as intended on the <p> element, giving it the specified styles.