Create textlink.html
This HTML document demonstrates the usage of the <base> element, which specifies the base URL for all relative URLs in the document. In this case, the base URL is set to https://www.google.com/, meaning that all relative links will be resolved from this base URL.

The document contains four links with different target attributes, specifying where the linked page should open:

Opens in New: The link opens in a new browser tab or window (target="_blank").

Opens in Self: The link opens in the same browsing context, i.e., the same window or tab (target="_self").

Opens in Parent: The link opens in the parent browsing context, which can be used if the page is embedded in a frame (target="_parent").

Opens in Body: The link opens in the full body of the browser window (target="_top"), replacing any frames.

Since the <base> tag is used, the relative URL /html/index.htm will be interpreted as https://www.google.com/html/index.htm.