Create adjacent sibling selector in internal css.html
In the provided HTML and CSS code:

The div element has a border of 4px solid.

The div + p selector targets the first <p> element that immediately follows a div element. This means:

The paragraph inside the div won't be blue because it is nested within the div.

The first <p> element after the div will be colored blue, as per the div + p rule.

The second paragraph after the div will not be blue, as it does not immediately follow the div element.

In summary:

The first paragraph (above the div) is unaffected and not blue.

The second paragraph (inside the div) is also not blue.

The third paragraph (immediately following the div) will be blue.

The fourth paragraph is not blue as it does not immediately follow the div.