Create pseudo class selector in internal css.html
In this HTML and CSS code:

a:hover:

This applies when the mouse pointer hovers over the anchor (<a>) element.

It changes the background color of the anchor to peachpuff, the text color to green, and increases the font size to 2em.

button:active:

This applies when the button is being clicked or activated (e.g., when it's being pressed down).

It changes the background color of the button to yellow.

Thus, when you hover over the "Google" link, the text will turn green, the background will become peachpuff, and the font will increase in size. When you click the button, its background will turn yellow.