Create undefined variable in internal js.html
This HTML code defines a JavaScript variable num without assigning it a value, which means it is undefined by default. When document.write() is used to display "The value of num is: " + num, it outputs: The value of num is: undefined, followed by a line break (<br/>) on the webpage.