Create null variable in internal js.html
This HTML code illustrates the use of null in JavaScript to represent an unknown or empty value. The variable houseNo is correctly declared and initialized to null. The variable apartment is declared with a string "B-2", but in the next line, there's a typo: appartment = null; (misspelled). This creates a new undeclared global variable appartment and leaves the original apartment unchanged. As a result, the output will be:
The value of the houseNo is: null
The value of the apartment is: B-2
demonstrating both the use of null and the importance of correct variable naming.