Create javascript object notation (internal js).html
This HTML code parses a JSON string to create a JavaScript object and displays a specific value from it. The JSON string text contains an object with an employees array, each element being an object with firstName and lastName. Using JSON.parse(), the string is converted to a JavaScript object stored in obj. The script then accesses the second employee in the array (index 1) and sets the content of the <p> element with ID demo to "Anna Smith", displaying that name on the webpage.