Create class creation in internal js.html
This HTML code demonstrates how to use the typeof operator in JavaScript to check the type of a class. The Car class is defined, but it has no body (no properties or methods). The typeof Car expression is used to determine the type of the Car class, which is a function in JavaScript. In JavaScript, classes are essentially syntactic sugar for constructor functions.