Create function parameters in internal js.html
This HTML code demonstrates the use of default parameters in JavaScript. The division function takes two parameters, a and b, with default values of 100 and 50, respectively. If no value is provided for a parameter when calling the function, the default value will be used.

The first call, division(10, 2), uses the provided values 10 for a and 2 for b, so the result is 10 / 2 = 5.

The second call, division(10, undefined), uses the provided 10 for a, but undefined for b triggers the default value for b (50), so the result is 10 / 50 = 0.2.

The third call, division(), uses the default values for both a and b (i.e., 100 and 50), so the result is 100 / 50 = 2.